pyproject.toml
[project]
name = "pesquisa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-use>=0.11.8",
    "matplotlib>=3.10.8",
    "openai>=2.15.0",
    "pandas>=3.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "statsmodels>=0.14.6",
]
